Refactoring FakeFindeks
Refactoring FakeFindeksService
