--- src/rentACar/Infrastructure/Adapters/FakeFindeksService/FakeFindeksServiceAdapter.cs
@@ -1,13 +1,0 @@
-﻿using Application.Services.FindeksService;
-
-namespace Infrastructure.Adapters.FakeFindeksService;
-
-public class FakeFindeksServiceAdapter : IFindeksService
-{
-    public short GetScore(string identityNumber)
-    {
-        Random random = new();
-        short score = Convert.ToInt16(random.Next(1900));
-        return score;
-    }
-}
